# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel Booking API",
        "description": "API endpoints for managing hotel bookings",
        "contact": {
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/rooms": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Get all rooms",
                "operationId": "getRooms",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RoomWithId"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Create a new room",
                "operationId": "createRoom",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoomPOST"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Room created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomWithId"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/rooms/{id}": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Get room by ID",
                "operationId": "getRoomById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of room to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomWithId"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Room not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Update a room",
                "operationId": "updateRoom",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of room to update",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoomPUT"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Room updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomWithId"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Room not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Delete a room",
                "operationId": "deleteRoom",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of room to delete",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Room deleted successfully"
                    },
                    "404": {
                        "description": "Room not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "RoomPOST": {
                "title": "Room for POST",
                "description": "Room object",
                "required": [
                    "number",
                    "type",
                    "price_per_night",
                    "status"
                ],
                "properties": {
                    "number": {
                        "description": "Number of the room",
                        "type": "integer",
                        "format": "int64",
                        "example": 101
                    },
                    "type": {
                        "description": "Type of the room",
                        "type": "string",
                        "example": "Single"
                    },
                    "price_per_night": {
                        "description": "Price per night for the room",
                        "type": "number",
                        "format": "double",
                        "example": 100
                    },
                    "status": {
                        "description": "Status of the room",
                        "type": "string",
                        "enum": [
                            "available",
                            "booked",
                            "maintenance"
                        ]
                    }
                },
                "type": "object"
            },
            "RoomPUT": {
                "title": "Room for PUT",
                "description": "Room object",
                "properties": {
                    "number": {
                        "description": "Number of the room",
                        "type": "integer",
                        "format": "int64",
                        "example": 101
                    },
                    "type": {
                        "description": "Type of the room",
                        "type": "string",
                        "example": "Single"
                    },
                    "price_per_night": {
                        "description": "Price per night for the room",
                        "type": "number",
                        "format": "double",
                        "example": 100
                    },
                    "status": {
                        "description": "Status of the room",
                        "type": "string",
                        "enum": [
                            "available",
                            "booked",
                            "maintenance"
                        ]
                    }
                },
                "type": "object"
            },
            "RoomWithId": {
                "title": "Room with ID",
                "description": "Room object",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the room",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "number": {
                        "description": "Number of the room",
                        "type": "integer",
                        "format": "int64",
                        "example": 101
                    },
                    "type": {
                        "description": "Type of the room",
                        "type": "string",
                        "example": "Single"
                    },
                    "price_per_night": {
                        "description": "Price per night for the room",
                        "type": "number",
                        "format": "double",
                        "example": 100
                    },
                    "status": {
                        "description": "Status of the room",
                        "type": "string",
                        "enum": [
                            "available",
                            "booked",
                            "maintenance"
                        ]
                    }
                },
                "type": "object"
            }
        }
    }
}